backend/api/requirements.txt
flask
flask-cors
tensorflow
tensorflow-hub
pillow
numpy
torch
transformers